Create KotlinDsl
When you run this program, it will prompt you to enter your name. After you do so and press Enter, it will print out a greeting to you. This is a basic example of a console application, but C# can be used to create much more complex programs, including web applications, desktop applications, and even games.